+++ components/popover/popover.stories.js
@@ -4,24 +4,44 @@ import {
   POPOVER_HORIZONTAL_ALIGNMENT,
   POPOVER_VERTICAL_ALIGNMENT,
   POPOVER_ROLES,
+  POPOVER_CONTENT_WIDTHS,
 } from './';
 import PopoverDefault from './popover_default.story.vue';
 import { createTemplateFromVueFile } from '../storybook_utils';
 import PopoverMdx from './popover.mdx';
+import { action } from '@storybook/addon-actions';
+import { TOOLTIP_HIDE_ON_CLICK_VARIANTS } from '../tooltip';
 
 const argTypesData = {
+  // Slots
+  anchor: {
+    table: {
+      type: {
+        summary: 'VNode',
+      },
+    },
+  },
+  content: {
+    control: 'text',
+    table: {
+      type: {
+        summary: 'VNode',
+      },
+    },
+  },
+
+  // Props
   id: {
-    defaultValue: 'DtPopover__content0',
     table: {
       defaultValue: {
-        summary: 'Automatically generated unique ID',
+        summary: 'generated unique ID',
       },
     },
   },
   padding: {
     control: {
       type: 'select',
-      options: POPOVER_PADDING_CLASSES,
+      options: Object.keys(POPOVER_PADDING_CLASSES),
     },
   },
   role: {
@@ -42,13 +62,50 @@ const argTypesData = {
       options: POPOVER_VERTICAL_ALIGNMENT,
     },
   },
+  contentWidth: {
+    control: {
+      type: 'select',
+      options: POPOVER_CONTENT_WIDTHS,
+    },
+  },
+
+  // Events
+  onClose: {
+    table: {
+      disable: true,
+    },
+  },
+
+  'update:open': {
+    description: `The popover will emit a "false" boolean value for this event when the \
+user performs a popover-closing action. Parent components can sync on this value to create \
+a 2-way binding to control popover visibility.`,
+    table: {
+      type: {
+        summary: 'boolean',
+      },
+    },
+  },
+  hideOnClick: {
+    type: 'select',
+    options: TOOLTIP_HIDE_ON_CLICK_VARIANTS,
+  },
+};
+
+// Default Props for all variations
+export const argsData = {
+  onClose: action('update:show'),
 };
 
 export default {
-  title: 'Elements/Popovers',
+  title: 'Components/Popovers',
   component: DtPopover,
+  args: argsData,
   argTypes: argTypesData,
   parameters: {
+    controls: {
+      sort: 'requiredFirst',
+    },
     docs: {
       page: PopoverMdx,
     },
@@ -106,6 +163,21 @@ FixedRight.parameters = {
 export const NoPadding = Template.bind({});
 NoPadding.args = { ...Default.args, padding: 'none', open: true };
 NoPadding.parameters = {
+  docs: {
+    source: {
+      code: `
+<dt-popover padding="none">
+  <template #anchor="{ attrs }"></template>
+  <template #content></template>
+</dt-popover>
+    `,
+    },
+  },
+};
+
+export const overlayModal = Template.bind({});
+overlayModal.args = { ...Default.args, open: true, modal: true };
+overlayModal.parameters = {
   docs: {
     source: {
       code: `