+++ components/tooltip/tooltip.stories.js
@@ -1,31 +1,30 @@
 import { createTemplateFromVueFile } from '../storybook_utils';
 import DtTooltip from './tooltip';
+import DtTooltipFlipTemplate from './tooltip_flip.story.vue';
+import DtTooltipDefault from './tooltip_default.story';
+import DtTooltipVariantsTemplate from './tooltip_variants';
+import { action } from '@storybook/addon-actions';
+
+import { TOOLTIP_DIRECTION_MODIFIERS, TOOLTIP_HIDE_ON_CLICK_VARIANTS } from './tooltip_constants';
 import DtTooltipMdx from './tooltip.mdx';
-import DtTooltipDefaultTemplate from './tooltip_default.story.vue';
-import DtTooltipVariantsTemplate from './tooltip_variants.story';
-import { TOOLTIP_DIRECTION_MODIFIERS } from './tooltip_constants';
 
 // Default Prop Values
 export const argsData = {
+  show: false,
   message: 'This is a Tooltip',
   anchor: 'Hover over me to see a tooltip',
   default: `This is a simple tooltip. The tooltip can be positioned in multiple areas too!`,
-  show: true,
+  onClose: action('update:show'),
 };
 
 export const argTypesData = {
-  // Props
-  show: {
-    description: 'Used to show tooltip',
-    control: 'boolean',
+  id: {
     table: {
-      category: 'props',
-      type: {
-        summary: 'boolean',
+      defaultValue: {
+        summary: 'generated unique ID',
       },
     },
   },
-
   arrowDirection: {
     control: {
       type: 'select',
@@ -33,12 +32,34 @@ export const argTypesData = {
     },
   },
 
+  flip: {
+    defaultValue: ['left-center', 'top-center'],
+  },
+
+  offset: {
+    defaultValue: [0, 0],
+  },
+
+  hideOnClick: {
+    type: 'select',
+    options: TOOLTIP_HIDE_ON_CLICK_VARIANTS,
+  },
+
+  flipBoundary: {
+    defaultValue: 'clippingParents',
+  },
+  transition: {
+    control: {
+      type: 'select',
+      options: ['', 'fade', 'slide-down', 'pop', 'shake'],
+    },
+  },
   // Slots
   default: {
     control: 'text',
     table: {
       type: {
-        summary: 'text/html',
+        summary: 'VNode',
       },
     },
   },
@@ -49,15 +70,31 @@ export const argTypesData = {
     table: {
       category: 'slots',
       type: {
-        summary: 'text/html',
+        summary: 'VNode',
       },
+    },
+  },
+  // Events
+  'update:show': {
+    description: `The tooltip will emit a "false" boolean value for this event when the \
+user performs a tooltip-closing action. Parent components can sync on this value to create \
+a 2-way binding to control tooltip visibility.`,
+    table: {
+      type: {
+        summary: 'boolean',
+      },
+    },
+  },
+  onClose: {
+    table: {
+      disable: true,
     },
   },
 };
 
 // Story Collection
 export default {
-  title: 'Elements/Tooltip',
+  title: 'Components/Tooltip',
   component: DtTooltip,
   args: argsData,
   argTypes: argTypesData,
@@ -70,12 +107,18 @@ export default {
 };
 
 // Templates
-const DefaultTemplate = (args, { argTypes }) => createTemplateFromVueFile(args, argTypes, DtTooltipDefaultTemplate);
-const VariantsTemplate = (args, { argTypes }) => createTemplateFromVueFile(args, argTypes, DtTooltipVariantsTemplate);
-
+const TooltipFlipTemplate = (args, { argTypes }) =>
+  createTemplateFromVueFile(args, argTypes, DtTooltipFlipTemplate);
+const TooltipDefaultTemplate = (args, { argTypes }) => createTemplateFromVueFile(args, argTypes, DtTooltipDefault);
+const TooltipVariantsTemplate = (args, { argTypes }) =>
+  createTemplateFromVueFile(args, argTypes, DtTooltipVariantsTemplate);
 // Stories
-export const Default = DefaultTemplate.bind({});
+
+export const Default = TooltipDefaultTemplate.bind({});
 Default.args = {};
 
-export const Variants = VariantsTemplate.bind({});
+export const Variants = TooltipVariantsTemplate.bind({});
 Variants.args = {};
+
+export const Flip = TooltipFlipTemplate.bind({});
+Flip.args = {};